chore(map): delete theme auto-detection and intensity dead code; move palettes to lib root (#341)
